+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 - fix `MmdStandardMaterial` not affected by light intensity
 
-- fix application of linked transform nodes being applied one frame late in `HumanoidMmd`
-
 - sphere map is now affected by light intensity
+
+- fix application of linked transform nodes being applied one frame late in `HumanoidMmd`
 
 ## 0.29.0